refactor(master): move VolumeAssigner to powerfs-allocator crate
P3 of allocator decoupling plan:
- Move VolumeAssigner trait + RoundRobin/ConsistentHash/Smart implementations
  from powerfs-master to powerfs-allocator
- Master's volume_assigner.rs now re-exports from powerfs_allocator
- All existing imports in master remain unchanged (backward compatible)
- 6 volume assigner tests pass (13 total in allocator crate)
